requirements.txt
flask==3.0.0
python-dotenv==1.0.0
duckduckgo-search==4.1.1
flask-wtf==1.2.1